+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.21-1.1.2] - 2024-12-21
+### Added
+- Glow berry juice, pie, tea, and wine
+- Chocolate bars, toasted cocoa beans (used to craft chocolate bars), chocolate milk, and hot cocoa
+
 ## [1.21-1.1.1] - 2024-09-15
 ### Changed
 - Barley block renamed to "Barley Straw Block" (ID is still `berriesandmore:barley_block`)